requirements.txt
# Web Framework
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6

# Database
sqlalchemy==2.0.23
alembic==1.12.1
psycopg[binary]==3.1.14
sqlalchemy-utils==0.41.1

# Data Validation
pydantic==2.5.0
pydantic-settings==2.1.0

# Authentication & Security
python-jose==3.3.0
passlib==1.7.4
bcrypt==4.1.1
python-dotenv==1.0.0

# API Documentation
python-multipart==0.0.6

# Caching & Message Queue
redis==5.0.1
celery==5.3.4



# Forecasting (later)
# prophet==1.1.5
# tensorflow==2.14.0

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.2

# Code Quality
black==23.12.0
flake8==6.1.0
isort==5.13.2

# Utilities
requests==2.31.0
python-dateutil==2.8.2
pytz==2023.3
